fix(nightly): clang-tidy narrowing + main exception-escape, and CPD apt-index refresh (#63)
Restores the nightly build to green (red since 2026-05-30) by fixing two
independent causes:

1. Whole-project clang-tidy: 7 narrowing-conversions from widening the
   statistics counters to int64_t (#27, #28, #30) made explicit, plus a
   top-level try/catch in main() resolving bugprone-exception-escape.
2. Copy-Paste Detection: apt-get update before installing default-jre-headless
   so a stale runner apt index no longer 404s on the openjdk patch version.

Verified on this branch: PR CI green and a manually-dispatched nightly run
passed all 5 jobs (Copy-Paste Detection and whole-project clang-tidy both
green under clang-tidy-19).

+++ src/core/statistics_manager.cpp
@@ -323,7 +323,7 @@ std::array<double, DIFFICULTY_COUNT> StatisticsManager::getCompletionRates() con
 
     for (size_t i = 0; i < DIFFICULTY_COUNT; ++i) {
         if (stats.games_played[i] > 0) {
-            rates[i] = static_cast<double>(stats.games_completed[i]) / stats.games_played[i];
+            rates[i] = static_cast<double>(stats.games_completed[i]) / static_cast<double>(stats.games_played[i]);
         } else {
             rates[i] = 0.0;
         }
@@ -580,7 +580,8 @@ void StatisticsManager::updateAggregateStats(const GameStats& completed_game) co
         // Update average rating
         auto current_rating_average = cached_stats_.average_ratings[diff_index];
         cached_stats_.average_ratings[diff_index] =
-            (current_rating_average * games_count_before + completed_game.puzzle_rating) / (games_count_before + 1);
+            (current_rating_average * static_cast<double>(games_count_before) + completed_game.puzzle_rating) /
+            static_cast<double>(games_count_before + 1);
     }
 
     // Update per-difficulty stats
